refactor: 관리자/사용자 엔티티 분리 및 인증 구조 개선
