+++ .github/workflows/deploy.yml
@@ -62,8 +62,9 @@ jobs:
               "SPRING_DATASOURCE_URL=${{ secrets.SPRING_DATASOURCE_URL }}",
               "SPRING_DATASOURCE_USERNAME=${{ secrets.SPRING_DATASOURCE_USERNAME }}",
               "SPRING_DATASOURCE_PASSWORD=${{ secrets.SPRING_DATASOURCE_PASSWORD }}",
-              "ADMIN_USER_ID=${{ secrets.ADMIN_USER_ID }}",
+              "ADMIN_LOGIN_ID=${{ secrets.ADMIN_LOGIN_ID }}",
               "ADMIN_PASSWORD=${{ secrets.ADMIN_PASSWORD }}",
+              "ADMIN_EMAIL=${{ secrets.ADMIN_EMAIL }}",
               "JWT_SECRET=${{ secrets.JWT_SECRET }}",
               "JWT_EXPIRATION_SECONDS=${{ secrets.JWT_EXPIRATION_SECONDS }}",
               "AWS_REGION=${{ secrets.AWS_REGION }}",
