--- src/main/java/com/example/cowmjucraft/domain/account/entity/Role.java
@@ -1,5 +1,0 @@
-package com.example.cowmjucraft.domain.account.entity;
-
-public enum Role {
-    ADMIN, USER
-}
