--- src/main/java/com/example/cowmjucraft/domain/account/entity/User.java
@@ -1,49 +1,0 @@
-package com.example.cowmjucraft.domain.account.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String userId;
-
-    @Column(nullable = false)
-    private String userName;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Role role;
-
-    public User(String userId, String userName, String email, String password, Role role) {
-        this.userId = userId;
-        this.userName = userName;
-        this.email = email;
-        this.password = password;
-        this.role = role;
-    }
-
-    public void updateUserId(String userId) {
-        this.userId = userId;
-    }
-
-    public void updatePassword(String password) {
-        this.password = password;
-    }
-}
