--- src/main/java/com/example/cowmjucraft/domain/auth/login/controller/AdminLoginController.java
@@ -1,42 +1,0 @@
-package com.example.cowmjucraft.domain.auth.login.controller;
-
-import com.example.cowmjucraft.domain.auth.login.dto.request.AdminLoginRequestDto;
-import com.example.cowmjucraft.domain.auth.login.dto.response.AdminLoginResponseDto;
-import com.example.cowmjucraft.domain.auth.login.service.AdminLoginService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseCookie;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.time.Duration;
-
-@Validated
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/api/admin")
-public class AdminLoginController implements AdminLoginControllerDocs {
-
-    private final AdminLoginService adminLoginService;
-
-    @Override
-    @PostMapping("/login")
-    public ResponseEntity<AdminLoginResponseDto> login(@RequestBody AdminLoginRequestDto request) {
-        AdminLoginService.LoginResult loginResult = adminLoginService.login(request);
-
-        ResponseCookie cookie = ResponseCookie.from("access_token", loginResult.token())
-                .httpOnly(true)
-                .secure(true)
-                .path("/")
-                .maxAge(Duration.ofSeconds(loginResult.expiresInSeconds()))
-                .build();
-
-        return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE, cookie.toString())
-                .body(loginResult.admin());
-    }
-}
