--- src/main/java/com/example/cowmjucraft/domain/auth/signup/service/UserService.java
@@ -1,8 +1,0 @@
-package com.example.cowmjucraft.domain.auth.signup.service;
-
-import com.example.cowmjucraft.domain.auth.signup.dto.response.SignupResponse;
-import com.example.cowmjucraft.domain.auth.signup.dto.request.SignupRequest;
-
-public interface UserService {
-    SignupResponse signup(SignupRequest request);
-}
