+++ src/main/java/com/example/cowmjucraft/domain/identity/admin/repository/AdminRepository.java
@@ -1,0 +1,11 @@
+package com.example.cowmjucraft.domain.identity.admin.repository;
+
+import com.example.cowmjucraft.domain.identity.admin.entity.Admin;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface AdminRepository extends JpaRepository<Admin, Long> {
+    Optional<Admin> findByLoginId(String loginId);
+    boolean existsByLoginId(String loginId);
+}
