+++ src/main/java/com/example/cowmjucraft/domain/identity/user/repository/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.example.cowmjucraft.domain.identity.user.repository;
+
+import com.example.cowmjucraft.domain.identity.user.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    boolean existsByUserId(String userId);
+    Optional<Member> findByUserId(String userId);
+}
