+++ src/main/java/com/example/cowmjucraft/domain/identity/user/signup/dto/request/UserSignupRequestDto.java
@@ -1,0 +1,12 @@
+package com.example.cowmjucraft.domain.identity.user.signup.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class UserSignupRequestDto {
+
+    private String userId;
+    private String password;
+    private String email;
+    private String userName;
+}
