+++ src/main/resources/application-local.yml
@@ -15,8 +15,9 @@ spring:
     show-sql: true
 
 admin:
-  user-id: ${ADMIN_USER_ID}
+  login-id: ${ADMIN_LOGIN_ID}
   password: ${ADMIN_PASSWORD}
+  email: ${ADMIN_EMAIL:}
 
 jwt:
   secret: ${JWT_SECRET}
